refactor: [RTP] code & file structure
